requirements.txt
mysql-connector-python
pyttsx3
python-dotenv
vosk   # when you add voice support
